refactor: hook 정리
