+++ frontend/src/hooks/index.ts
@@ -1,7 +1,6 @@
-export { default as useMain } from './useMain';
 export { default as useQuizSetting } from './useQuizSetting';
 export { default as useQuiz } from './useQuiz';
-export { default as useCards } from './useCards';
+export { default as useCard } from './useCard';
 export { default as useSnackbar } from './useSnackbar';
 export { default as useTimeout } from './useTimeout';
 export { default as useRouter } from './useRouter';
