+++ frontend/src/recoil/index.ts
@@ -5,6 +5,7 @@ export {
   workbookIdState,
   publicWorkbookIdState,
   workbookState,
+  editedWorkbookState,
 } from './workbookState';
 export {
   cardIdState,
